fix: fix transfer bug
