+++ examples/contracts/erc20/erc20.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/bcl-chain/web3.go/mobile"
+	web3go "github.com/bcl-chain/web3.go/mobile"
 )
 
 func main() {
@@ -16,22 +16,22 @@ func main() {
 	fmt.Println("we have a connection")
 
 	//实例化合约
-	address, _ := web3go.NewAddressFromHex("0xc4297a75cded5d8e33ec571c79351895c9b362df")
+	address, _ := web3go.NewAddressFromHex("0xC2870160634934c79ac93757E37b72AE17141715")
 	erc20, err := web3go.NewERC20(address, client)
 	if err != nil {
 		log.Fatal(err)
 	}
 
 	//打印用户token余额
-	address, _ = web3go.NewAddressFromHex("0xfe04cb1d7d6715169edc07c8e3c2fdba3a0854af")
+	address, _ = web3go.NewAddressFromHex("0x941c69B23CeF5f5021b5966f4ba85fE6Bf9A58E1")
 	balance, err := erc20.BalanceOf(address)
 	if err != nil {
 		log.Fatal(err)
 	}
 	fmt.Println(balance.String())
 
-    //签名某交易
-	opts := web3go.NewTransactOpts("ac5c70f5a1c9f5b121edbbabb054d8763ac1f5066aeacbb14a5e4c10216fb605")
+	//签名某交易
+	opts := web3go.NewTransactOpts("72149D7D1FAD0A66AF39AB1693F28E66E0600A2032D874FF321C8ABEBDC41670")
 	nonce, err := client.GetPendingNonceAt(web3go.NewContext(), opts.GetFrom())
 	if err != nil {
 		log.Fatal(err)
@@ -44,14 +44,18 @@ func main() {
 	opts.SetGasLimit(300000)
 	opts.SetGasPrice(gasPrice)
 
-    //从某账户地址发起交易,打印交易签名
-	address, _ = web3go.NewAddressFromHex("0xf80af2b63ee5e02a53e28c3ed470d8dd54a068d6")
-	tx, _ := erc20.BuildTransfer(opts, address, web3go.NewBigInt(10))
+	//从某账户地址发起交易,打印交易签名
+	address, _ = web3go.NewAddressFromHex("0x71ce870012d1B6cd6e94E764e03643048617cA0F")
+	tx, _ := erc20.BuildTransfer(opts, address, "1500", 18, 8888)
 	fmt.Println(tx.GetHash().GetHex())
-	tx, _ = erc20.Transfer(opts, address, web3go.NewBigInt(10))
-	fmt.Println(tx.GetHash().GetHex())
+	// tx, _ = erc20.Transfer(opts, address, web3go.NewBigInt(10))
+	// fmt.Println(tx.GetHash().GetHex())
+	err = erc20.SendTransfer(tx)
+	if err != nil {
+		log.Fatal(err)
+	}
 
-    //打印账户token余额
+	//打印账户token余额
 	address, _ = web3go.NewAddressFromHex("0xfe04cb1d7d6715169edc07c8e3c2fdba3a0854af")
 	balance, err = erc20.BalanceOf(address)
 	if err != nil {
